[#154747783] Renamed propMapper
