+++ app/javascript/components/applications/flagged/FlaggedApplicationsIndexPage.js
@@ -3,7 +3,7 @@ import React from 'react'
 import IndexTable from '../../IndexTable'
 import TableLayout from '../../layouts/TableLayout'
 import mapProps from '~/utils/mapProps'
-import { mapFlaggedRecords } from '~/components/propMappers'
+import { mapFlaggedRecords } from '~/components/mappers/soqlToDomain'
 
 const tableFields = {
   "id":null,
