+++ app/javascript/components/listings/ListingApplicationsPage.js
@@ -3,7 +3,7 @@ import React from 'react'
 import IndexTable from '../IndexTable'
 import TableLayout from '../layouts/TableLayout'
 import mapProps from '~/utils/mapProps'
-import { mapListingApplicationPage, mapListing } from '~/components/propMappers'
+import { mapListingApplicationPage, mapListing } from '~/components/mappers/soqlToDomain'
 
 const tableFields = {
   "id":null,
