+++ app/javascript/components/listings/ListingPage.js
@@ -3,7 +3,7 @@ import React from 'react'
 import ListingDetails from './ListingDetails'
 import CardLayout from '../layouts/CardLayout'
 import mapProps from '~/utils/mapProps'
-import { mapListingDetails } from '~/components/propMappers'
+import { mapListingDetails } from '~/components/mappers/soqlToDomain'
 
 const ListingPageDetails = ({ listing }) => {
   return <ListingDetails listing={listing} />
