+++ app/javascript/components/listings/ListingsPage.js
@@ -3,7 +3,7 @@ import React from 'react'
 import IndexTable from '../IndexTable'
 import TableLayout from '../layouts/TableLayout'
 import mapProps from '~/utils/mapProps'
-import { mapListingsTableItem } from '~/components/propMappers'
+import { mapListingsTableItem } from '~/components/mappers/soqlToDomain'
 
 const tableFields = {
   "id":null,
