+++ app/javascript/components/mappers/soqlToDomain.js
